pyproject.toml
[tool.poetry]
name = "sep-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
django-ninja = "^1.3.0"
psycopg2 = "^2.9.10"
django-environ = "^0.12.0"
punq = "^0.7.0"
django-cors-headers = "^4.7.0"
bcrypt = "^4.3.0"
pyjwt = "^2.10.1"
factory-boy = "^3.3.3"
pytest-django = "^4.11.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.project.settings"